+++ src/containerElement.ts
@@ -3,7 +3,7 @@
  * @see https://www.w3.org/TR/SVG11/intro.html#TermContainerElement
  */
 
-import { ParsedElement } from "./domParser";
+import { ParsedElement } from "./svgParser";
 import { traverse } from "./traverse";
 
 export function collectContainer(pe: ParsedElement): string[] {
