+++ src/nodeMode.ts
@@ -2,7 +2,7 @@ import { refleshContent, svgRealMap, sendBackToEditor, contentChildrenComponent,
 import { SvgTag } from "./svg";
 import { svgPathManager } from "./pathHelpers";
 import { Vec2, v, vfp, OneOrMore } from "./utils";
-import { PathCommand, ParsedElement } from "./domParser";
+import { PathCommand, ParsedElement } from "./svgParser";
 import { Mode } from "./abstractMode";
 import { applyToPoint, inverse, toString } from "transformation-matrix";
 import { shaper } from "./shapes";
