+++ src/svgParser.ts
